refactor: move OAuth refresh to shared util; add pre-spawn refresh in container-runner
- Extract core refresh logic to src/oauth-token-refresh.ts so both the
  health-monitor and the container spawner can use the same code
- health-monitor/token-refresh.ts is now a thin wrapper around the shared util
- In spawnContainer(), refresh the token before buildMounts() if it's
  expiring within 60 min — fixes the shutdown case where a token expires
  while the host is off and a task fires immediately on boot before the
  health-monitor's 5-min check has a chance to run
- Also fixes the remaining main-branch copy of token-refresh.ts which still
  had the old form-encoded body (Cloudflare 403); shared util uses JSON
